+++ components/server/src/common.rs
@@ -1,25 +1,69 @@
 // Copyright 2023 TiKV Project Authors. Licensed under Apache-2.0.
 use std::{
-    cmp,
+    cmp, env,
+    net::SocketAddr,
     path::{Path, PathBuf},
-    sync::Arc,
+    sync::{mpsc, Arc},
     u64,
 };
 
 use encryption_export::{data_key_manager_from_config, DataKeyManager};
+use engine_rocks::FlowInfo;
 use error_code::ErrorCodeExt;
-use file_system::File;
+use file_system::{set_io_rate_limiter, BytesFetcher, File};
 use tikv::config::TikvConfig;
 use tikv_util::sys::{disk, path_in_diff_mount_point};
 
+/// This is the common layer of TiKV-like servers. By holding it in its own
+/// TikvServer implementation, one can easily access the common ability of a
+/// TiKV server.
 pub struct TikvServerCore {
     pub config: TikvConfig,
     pub store_path: PathBuf,
     pub lock_files: Vec<File>,
     pub encryption_key_manager: Option<Arc<DataKeyManager>>,
+    pub flow_info_sender: Option<mpsc::Sender<FlowInfo>>,
+    pub flow_info_receiver: Option<mpsc::Receiver<FlowInfo>>,
 }
 
 impl TikvServerCore {
+    pub fn check_conflict_addr(&mut self) {
+        let cur_addr: SocketAddr = self
+            .config
+            .server
+            .addr
+            .parse()
+            .expect("failed to parse into a socket address");
+        let cur_ip = cur_addr.ip();
+        let cur_port = cur_addr.port();
+        let lock_dir = get_lock_dir();
+
+        let search_base = env::temp_dir().join(lock_dir);
+        file_system::create_dir_all(&search_base)
+            .unwrap_or_else(|_| panic!("create {} failed", search_base.display()));
+
+        for entry in file_system::read_dir(&search_base).unwrap().flatten() {
+            if !entry.file_type().unwrap().is_file() {
+                continue;
+            }
+            let file_path = entry.path();
+            let file_name = file_path.file_name().unwrap().to_str().unwrap();
+            if let Ok(addr) = file_name.replace('_', ":").parse::<SocketAddr>() {
+                let ip = addr.ip();
+                let port = addr.port();
+                if cur_port == port
+                    && (cur_ip == ip || cur_ip.is_unspecified() || ip.is_unspecified())
+                {
+                    let _ = try_lock_conflict_addr(file_path);
+                }
+            }
+        }
+
+        let cur_path = search_base.join(cur_addr.to_string().replace(':', "_"));
+        let cur_file = try_lock_conflict_addr(cur_path);
+        self.lock_files.push(cur_file);
+    }
+
     pub fn init_fs(&mut self) {
         let lock_path = self.store_path.join(Path::new("LOCK"));
 
@@ -135,4 +179,61 @@ impl TikvServerCore {
         .unwrap()
         .map(Arc::new);
     }
+
+    pub fn init_io_utility(&mut self) -> BytesFetcher {
+        let stats_collector_enabled = file_system::init_io_stats_collector()
+            .map_err(|e| warn!("failed to init I/O stats collector: {}", e))
+            .is_ok();
+
+        let limiter = Arc::new(
+            self.config
+                .storage
+                .io_rate_limit
+                .build(!stats_collector_enabled /* enable_statistics */),
+        );
+        let fetcher = if stats_collector_enabled {
+            BytesFetcher::FromIoStatsCollector()
+        } else {
+            BytesFetcher::FromRateLimiter(limiter.statistics().unwrap())
+        };
+        // Set up IO limiter even when rate limit is disabled, so that rate limits can
+        // be dynamically applied later on.
+        set_io_rate_limiter(Some(limiter));
+        fetcher
+    }
+
+    pub fn init_flow_receiver(&mut self) -> engine_rocks::FlowListener {
+        let (tx, rx) = mpsc::channel();
+        self.flow_info_sender = Some(tx.clone());
+        self.flow_info_receiver = Some(rx);
+        engine_rocks::FlowListener::new(tx)
+    }
+}
+
+#[cfg(unix)]
+fn get_lock_dir() -> String {
+    format!("{}_TIKV_LOCK_FILES", unsafe { libc::getuid() })
+}
+
+#[cfg(not(unix))]
+fn get_lock_dir() -> String {
+    "TIKV_LOCK_FILES".to_owned()
+}
+
+fn try_lock_conflict_addr<P: AsRef<Path>>(path: P) -> File {
+    let f = File::create(path.as_ref()).unwrap_or_else(|e| {
+        fatal!(
+            "failed to create lock at {}: {}",
+            path.as_ref().display(),
+            e
+        )
+    });
+
+    if f.try_lock_exclusive().is_err() {
+        fatal!(
+            "{} already in use, maybe another instance is binding with this address.",
+            path.as_ref().file_name().unwrap().to_str().unwrap()
+        );
+    }
+    f
 }
